github/requirements.txt
selenium
webdriver-manager
sentence-transformers
spacy
nltk
torch
transformers
huggingface-hub
pandas
praw
python-dotenv
PyGithub
python-dotenv

# python -m spacy download en_core_web_sm